requirements.txt
matplotlib~=3.4.2
scikit-learn~=0.24.2
visdom~=0.1.8.9
numpy~=1.20.3
pytorch~=1.9.1